Inital commit of PyRSoXS ========================
This is the initial commit of PyRSoXS.  It contains the following:

* PyRSoXS - the main package
* core - the internal package for all core data structure and functions
* _rotated_geometry.py - the main module for the rotated geometry

As of now, the package is not installable. It is just a collection of
files. Future commits will add the necessary files to make it installable.

As of now, the package does not reduce the data and has no integration
with igor pro. Future commits will add this functionality.